Update document.md
A Glaucoma Detection using Convolutional Neural
Network


Glaucoma is the disease that affects the human eye and results in a permanent blindness if not detect at early stage. The Situation is very complex, so the proper detection is must. Detection at early stages may be improved else it may lead to loss of vision. The main reason for the vision loss is affected optic nerve of the eye. The doctor needs at least 5 check-up reports to confirm the glaucoma a ffection, so it becomes essential to design a system to detect this disease accurately in one attempt. 

This paper presents architecture for the proper glaucoma detection based on the  convolutional neural network (CNN). It will make differentiation between the patterns formed for glaucoma and non-glaucoma. CNN used for achieving the adequate  performance in the glaucoma detection. Overall Architecture is having six layers for proper detection of the disease. A dropout mechanism is also applied to improve the performance of the given approach. SCES(has 46 images for glaucoma and 1676 images for the normal fundus.) and ORIGA(having 168 glaucoma and 482 images of the normal fundus) datasets used for the experiments and obtained values are .822 and .882 for the ORIGA and SCES respectively. The major objective is to find the most  similar patterns in between the normal human eye and the infected glaucoma eye.

The presented approach works in six layers. The four layers are the convolutional layers, and the last two layers are fully connected. Presented CNN to get the input, the ROI of the image  is used instead of full image for the Faster processing compared to full disc or cup images and Improved detection speed for glaucoma.
Also used ARGA LI Approach (Adaptive Region Growing and Level Set Integration) for the removal of bright fringe to improve image quality also helps determine the center and radius for trimming as well as fixed to 256 x 256 pixels for consistency.
Dropout Implemented in two fully connected layers. To prevent overfitting, data augmentation generates translations and horizontal reflections of images. The area under the curve () AUC is utilised of the receiver operating characteristics (ROC) curve for the evaluation of glaucoma detecting performance of the model. 

For validating the proposed CNN-based approach for glaucoma detection, it was compared with state-of-the-art reconstruction-based methods. The model was trained on 90 images from a 600-image dataset, and the remaining images were used for testing. The proposed method achieved an AUC of 0.822 for the ORIGA dataset and 0.882 for the SCES dataset, The state-of-the-art method, achieved an AUCs of 0.809 and 0.859, respectively. The detection capability of the proposed system seems higher than other methods and effective for glaucoma detection.





                                                                                                                                                     Naz Mahwish